Sort py files; Merge similar folders
