+++ c++/DP Basics Using CPP/check_prime.cpp
