+++ c++/DP Basics Using CPP/fence_paint.cpp
