+++ python/PseudoRandomSeriesEncrypter.py
